test: add tag tests and separate tag arbitraries
