+++ test/html/create-element/arbitraries.js
@@ -1,16 +1,6 @@
 import { fc } from 'ava-fast-check'
 import { createElement } from 'html/create-element'
-import {
-  nonSelfClosingTags,
-  tags,
-  selfClosingTags
-} from 'html/create-element/tags'
-
-export const tagArb = () => fc.constantFrom(...tags)
-
-export const selfClosingTagArb = () => fc.constantFrom(...selfClosingTags)
-
-export const nonSelfClosingTagArb = () => fc.constantFrom(...nonSelfClosingTags)
+import { nonSelfClosingTagArb, selfClosingTagArb } from './tags/arbitraries'
 
 export const attributeNameArb = fc.string
